Remove all broken `of` methods
The `of` method is meant to be used on the data type, for example:

    Option.of

It is not meant to be used on each tag:

    Option.left.of
    Option.right.of

The above doesn't make sense. The tags should be plain functions.

The `of` method only takes a single parameter:

    Option.of(1)

It is not meant to take more than one:

    Option.of(1, 2)

Fantasy Land leaves multiple arguments as unspecified (though, I
strongly believe bilby.js should make it throw an exception) but
anything that *requires* more than one argument is definitely broken.